+++ src/main/java/sayTheSpire/config/Setting.java
@@ -1,0 +1,52 @@
+package sayTheSpire.config;
+
+import com.google.gson.JsonElement;
+
+import sayTheSpire.Output;
+import sayTheSpire.localization.LocalizationContext;
+
+public abstract class Setting {
+
+    protected LocalizationContext localization;
+    private SettingCategory parent;
+    private String name;
+
+    public Setting(SettingCategory parent, String name) {
+        this.parent = parent;
+        this.name = name;
+        this.localization = Output.localization.getContext(this.getLocalizationPath());
+    }
+
+    public abstract void fromJsonElement(JsonElement json);
+
+    public abstract JsonElement toJsonElement();
+
+    public abstract Object getValue();
+
+    public abstract void setValue(Object value);
+
+    public String getDescription() {
+        return this.localization.localize("description");
+    }
+
+    public String getLabel() {
+        return this.localization.localize("label");
+    }
+
+    public String getLocalizationPath() {
+        SettingCategory parent = this.getParent();
+        if (parent == null) {
+            return ".config.settings";
+        }
+        return parent.getLocalizationPath() + ".settings." + this.getName();
+    }
+
+    public String getName() {
+        return this.name;
+    }
+
+    public SettingCategory getParent() {
+        return this.parent;
+    }
+
+}