+++ src/main/java/sayTheSpire/ui/dynamic/events/ToggleChangeEvent.java
@@ -4,7 +4,7 @@ import sayTheSpire.ui.dynamic.elements.DynamicElement;
 
 public class ToggleChangeEvent extends DynamicEvent {
 
-    public Boolean value;
+    public final Boolean value;
 
     public ToggleChangeEvent(DynamicElement target, Boolean value) {
         super("toggleChange", target);